banco de dados na nuvem

+++ ProjetoFinal/Data/VO/BookVO.cs
@@ -4,7 +4,7 @@
     {
         public string Isbn { set; get; }
         public string Category { set; get; }
-        public float Price { set; get; }
+        public decimal Price { set; get; }
         public long Id { set; get; }
 
     }
